Live project quality updates, new MOTD and money limit increase

+++ api/models.nim
@@ -53,7 +53,7 @@ type
   Player* {.tableName: "Players".} = ref object of Model
     code*: PaddedStringOfCap[8] = newPaddedStringOfCap[8]("")
     authToken*: Option[PaddedStringOfCap[12]] = none PaddedStringOfCap[12]
-    money*: int32 = 0
+    money*: int64 = 0
     gender*: PaddedStringOfCap[1] = newPaddedStringOfCap[1]("M")
     firstname*: int16 = 0
     lastname*: int16 = 0
